Changed messagetype DoublesToRos and serialrunningport
git-svn-id: https://roboticssrv.wtb.tue.nl/svn/ros/code/tue-ros-pkg/trunk/rtt_generic_components@21281 3a3c2cfb-d3b9-409e-9e75-db8fcdbcaac6

+++ rtt_ROS_interfacing/src/DoublesToROS.cpp
@@ -51,7 +51,7 @@ void DoublesToROS::updateHook()
 	{
 		for ( uint i = 0; i < Ndouble; i++ )
 		{
-			std_msgs::Float64 msg;
+			std_msgs::Float32 msg;
 			msg.data = values[i];
 			doubleoutports[i].write( msg );
 		}
